build fixed for shared
