+++ packages/api/src/clear.ts
@@ -1,10 +1,10 @@
 import type { FastifyInstance } from 'fastify';
-import { clean } from '@lawbrador/shared/src/db/utils';
-import {ALL_ENDPOINT} from '@lawbrador/shared/src/endpoints';
+import {Endpoints, clean} from '@lawbrador/shared';
 
 export default async (fastify: FastifyInstance) => {
-	fastify.delete(ALL_ENDPOINT, {}, async () => {
+	fastify.delete(Endpoints.ALL, {}, async () => {
 		await clean(fastify.client);
 		return { status: 'success' };
 	});
 };
+
