+++ packages/api/src/get_or_scrape_document.ts
@@ -1,26 +1,25 @@
-import { applyConfig } from "@lawbrador/shared/src/rule_applyer";
-import {
-  RestructuredDocument,
-  restructuredDocument,
-  ScrapeRequest,
-  scrapeRequest
-} from "@lawbrador/shared/src/schemas/scrape";
-import { GET_OR_SCRAPE_DOCUMENT } from "@lawbrador/shared/src/endpoints";
 import { FastifyInstance } from "fastify";
-import { scrape } from "@lawbrador/shared/src/scraper";
-import { extractUrlVariables } from "@lawbrador/shared/src/url";
-import {createHash} from "./utils";
+import {
+  extractUrlVariables,
+  Endpoints,
+  Schemas,
+  restructure,
+  RestructuredDocument,
+  scrape,
+  ScrapeRequest,
+} from "@lawbrador/shared";
+import { createHash } from "./utils";
 
 export default async (fastify: FastifyInstance) => {
   fastify.get<{ Querystring: ScrapeRequest }>(
-    GET_OR_SCRAPE_DOCUMENT, 
+    Endpoints.GET_OR_SCRAPE_DOCUMENT,
     {
-	schema: {
-		querystring: scrapeRequest,
-		response: {
-			200:restructuredDocument 
-		}
-	}
+      schema: {
+        querystring: Schemas.scrapeRequest,
+        response: {
+          200: Schemas.restructuredDocument,
+        },
+      },
     },
     async (req): Promise<RestructuredDocument> => {
       const url = req.query.url;
@@ -29,14 +28,14 @@ export default async (fastify: FastifyInstance) => {
       );
       const hash = createHash(
         extractUrlVariables(url, config.documentUrlConfig),
-	config
+        config
       );
       const documentScrape =
         await fastify.collections.documentScrapeCache.cachedOrCreated(
           { hash: hash },
           async () => await scrape(url, hash)
         );
-      return applyConfig(documentScrape, config);
+      return restructure(documentScrape, config);
     }
   );
 };
