+++ packages/api/src/utils.ts
@@ -1,7 +1,6 @@
 import { TSchema } from '@sinclair/typebox';
 import { RouteShorthandOptions } from 'fastify';
-import {SourceSiteConfig} from '@lawbrador/shared/src/schemas/rules';
-import { hashObject } from "@lawbrador/shared/src/url";
+import { hashObject, SourceSiteConfig } from "@lawbrador/shared";
 
 export function routeConfig(response: TSchema, request: TSchema = null): RouteShorthandOptions {
 	const options: RouteShorthandOptions = { schema: {} };
