+++ packages/shared/src/constants/endpoints.ts
@@ -1,5 +1,5 @@
 export const API_PREFIX = "/api";
-export const SOURCES_ENDPOINT = `${API_PREFIX}/sources`;
-export const SEARCH_ENDPOINT = `${API_PREFIX}/search`;
-export const ALL_ENDPOINT = `${API_PREFIX}/all`;
+export const SOURCES = `${API_PREFIX}/sources`;
+export const SEARCH = `${API_PREFIX}/search`;
+export const ALL = `${API_PREFIX}/all`;
 export const GET_OR_SCRAPE_DOCUMENT = `${API_PREFIX}/scrape/document`
