+++ packages/shared/src/db/Collection.ts
@@ -1,6 +1,5 @@
 import {Collection, MongoClient, ObjectId, Document, ClientSession, } from "mongodb";
-import { DB_NAME }  from './constants';
-import type { MongoIdentity, Identity } from '../schemas/generic';
+import { DB_NAME, Identity, MongoIdentity }  from '..';
 
 function objectIdToString<T extends MongoIdentity>(obj: T): T & Identity {
 	return { ...obj, _id: obj._id.toString() };
