+++ packages/shared/src/restructure/index.ts
@@ -1,8 +1,10 @@
-import {RestructuredDocument, ScrapeResult, SourceSiteConfig} from "..";
+import { RestructuredDocument, ScrapeResult, SourceSiteConfig } from "..";
+import { applyRuleSet } from "./applyRuleSet";
+import { selectRuleSet } from "./selectRuleSet";
 
 export function restructure(
   scrapeResult: ScrapeResult,
-  config: SourceSiteConfig 
+  config: SourceSiteConfig
 ): RestructuredDocument {
   const ruleSet = selectRuleSet(scrapeResult.body, config);
   const body = applyRuleSet(scrapeResult.body, ruleSet);
