+++ packages/shared/src/schemas/generic.ts
@@ -1,34 +1,19 @@
-import { Static, Type } from '@sinclair/typebox';
-import type { ObjectId } from 'mongodb';
+import {  Type } from '@sinclair/typebox';
 
 export const identity = Type.Object({
 	_id: Type.String(),
 });
 
-export type Identity = Static<typeof identity>;
-
-export interface MongoIdentity {
-	_id:  ObjectId;
-}
-export interface LawbradorEvent<T> {
-	type: string;
-	data: T;
-}
-
 export const hashed = Type.Object({
 	hash: Type.String()
 })
-
-export type Hashed = Static<typeof hashed>;
 
 export enum ResponseStatus {
 	success,
 	failed,
 }
 
-
 export const statusResponse = Type.Object({
 	status: Type.Enum(ResponseStatus),
 });
 
-export type StatusResponse = Static<typeof statusResponse>;
