+++ packages/shared/src/schemas/genericTypes.ts
@@ -1,0 +1,14 @@
+import { identity, hashed, statusResponse } from "./generic";
+import { Static } from "@sinclair/typebox";
+import type { ObjectId } from "mongodb";
+
+export type Hashed = Static<typeof hashed>;
+export type StatusResponse = Static<typeof statusResponse>;
+export type Identity = Static<typeof identity>;
+export interface MongoIdentity {
+  _id: ObjectId;
+}
+export interface LawbradorEvent<T> {
+  type: string;
+  data: T;
+}
