+++ packages/shared/src/schemas/ruleConstants.ts
@@ -1,0 +1,3 @@
+
+export const ALL_SELECTION_OPERATORS = ["is", "includes", "regex"] as const;
+export const ALL_SELECTION_LOCATIONS = ["tag", "class", "id", "link", "text"] as const;
