+++ packages/shared/src/schemas/rules.ts
@@ -1,35 +1,28 @@
-import { Static, Type } from "@sinclair/typebox";
-import {markupNotation} from "./tags";
-
-export const ALL_SELECTION_OPERATORS = ["is", "includes", "regex"] as const;
+import { Type } from "@sinclair/typebox";
+import { markupNotation } from "./tags";
+import {
+  ALL_SELECTION_LOCATIONS,
+  ALL_SELECTION_OPERATORS,
+} from "./ruleConstants";
 
 export const selectionOperator = Type.Union(
-  ALL_SELECTION_OPERATORS.map((op) => Type.Literal(op)),
+  ALL_SELECTION_OPERATORS.map((op) => Type.Literal(op))
 );
-
-export type SelectionOperator = typeof ALL_SELECTION_OPERATORS[number];
-
-export const ALL_SELECTION_LOCATIONS = ["tag", "class", "id", "link", "text"] as const;
 
 export const selectionLocation = Type.Union(
-  ALL_SELECTION_LOCATIONS.map((loc) => Type.Literal(loc)),
+  ALL_SELECTION_LOCATIONS.map((loc) => Type.Literal(loc))
 );
-
-export type SelectionLocation = typeof ALL_SELECTION_LOCATIONS[number];
 
 export const selectionRule = Type.Object({
   op: selectionOperator,
   location: selectionLocation,
   value: Type.String({ minLength: 1 }),
 });
-export type SelectionRule = Static<typeof selectionRule>;
 
 export const markupRule = Type.Object({
-	tag: markupNotation,
-	filter: selectionRule
-})
-
-export type MarkupRule = Static<typeof markupRule>;
+  tag: markupNotation,
+  filter: selectionRule,
+});
 
 export const documentRuleSet = Type.Object({
   name: Type.Optional(Type.String()),
@@ -38,8 +31,6 @@ export const documentRuleSet = Type.Object({
   bodyRule: Type.Optional(selectionRule),
 });
 
-export type DocumentRuleSet = Static<typeof documentRuleSet>;
-
 export const htmlSearchRuleSet = Type.Object({
   pageVariable: Type.String({ minLength: 1 }),
   queryVariable: Type.String({ minLength: 1 }),
@@ -47,47 +38,37 @@ export const htmlSearchRuleSet = Type.Object({
   resultLinkRule: Type.Optional(selectionRule),
 });
 
-export type HtmlSearchRuleSet = Static<typeof htmlSearchRuleSet>;
-
 export const valueWithDisplayName = Type.Object({
   value: Type.String({ minLength: 1 }),
   displayName: Type.Optional(Type.String()),
 });
 
-export type ValueWithDisplayName = Static<typeof valueWithDisplayName>;
-
-export const staticUrlComponent = Type.Object({ value: Type.String({ minLength: 1 }) });
-
-export type StaticUrlComponent = Static<typeof staticUrlComponent>;
+export const staticUrlComponent = Type.Object({
+  value: Type.String({ minLength: 1 }),
+});
 
 export const variableUrlComponent = Type.Object({
   variableName: Type.String({ minLength: 1 }),
-  possibleValues: Type.Optional(Type.Array(valueWithDisplayName, { minItems: 1, uniqueItems: true })),
+  possibleValues: Type.Optional(
+    Type.Array(valueWithDisplayName, { minItems: 1, uniqueItems: true })
+  ),
 });
-
-export type VariableUrlComponent = Static<typeof variableUrlComponent>;
 
 export const urlComponent = Type.Union([
   staticUrlComponent,
   variableUrlComponent,
 ]);
 
-export type UrlComponent = Static<typeof urlComponent>;
-
 export const queryParam = Type.Object({
   name: Type.String({ minLength: 1 }),
   urlComponent: urlComponent,
 });
-
-export type QueryParam = Static<typeof queryParam>;
 
 export const urlConfig = Type.Object({
   base: Type.String({ format: "uri" }),
   pathComponents: Type.Array(urlComponent),
   queryComponents: Type.Array(queryParam),
 });
-
-export type UrlConfig = Static<typeof urlConfig>;
 
 export const sourceSiteConfig = Type.Object({
   _id: Type.String(),
@@ -100,6 +81,4 @@ export const sourceSiteConfig = Type.Object({
 
 export const sourceSiteConfigs = Type.Array(sourceSiteConfig);
 
-export type SourceSiteConfig = Static<typeof sourceSiteConfig>;
-
-export const unsavedSourceSiteConfig = Type.Omit(sourceSiteConfig, ['_id'])
+export const unsavedSourceSiteConfig = Type.Omit(sourceSiteConfig, ["_id"]);
