+++ packages/shared/src/schemas/schemaIndex.ts
@@ -1,0 +1,5 @@
+export * from './generic';
+export * from './rules';
+export * from './scrape';
+export * from './search';
+export * from './tags';
