+++ packages/shared/src/schemas/search.ts
@@ -1,4 +1,4 @@
-import { Static, Type } from '@sinclair/typebox';
+import {  Type } from '@sinclair/typebox';
 
 export const searchResult = Type.Object({
 	text: Type.String(),
@@ -6,17 +6,14 @@ export const searchResult = Type.Object({
 	hash: Type.String()
 });
 
-export type SearchResult = Static<typeof searchResult>;
 
 export const searchResponse = Type.Object({
 	results: Type.Array(searchResult),
 })
 
-export type SearchResponse = Static<typeof searchResponse>;
 
 export const searchRequest = Type.Object({
 	sourceConfigId: Type.String(),
 	searchParams: Type.Record(Type.String(), Type.String()),
 });
 
-export type SearchRequest = Static<typeof searchRequest>;
