+++ packages/shared/src/schemas/searchTypes.ts
@@ -1,0 +1,5 @@
+import {searchRequest, searchResponse, searchResult} from './search';
+import { Static } from '@sinclair/typebox';
+export type SearchResult = Static<typeof searchResult>;
+export type SearchResponse = Static<typeof searchResponse>;
+export type SearchRequest = Static<typeof searchRequest>;
