+++ packages/shared/src/utils/utils.ts
@@ -1,4 +1,4 @@
-import {Identity} from "./schemas/generic";
+import {Identity} from "..";
 
 export function last(array: any[]) {
 	return array[array.length - 1];
